🚚 move operate on store into own file
