+++ packages/atom.io/src/internal/set-state/set-into-store.ts
@@ -1,16 +1,9 @@
 import type { WritableFamilyToken, WritableToken } from "atom.io"
-import { type Canonical, parseJson } from "atom.io/json"
+import type { Canonical } from "atom.io/json"
 
-import { seekInStore } from "../families"
-import { getFamilyOfToken } from "../families/get-family-of-token"
-import { mintInStore } from "../families/mint-in-store"
-import type { OpenOperation } from "../operation"
-import { closeOperation, openOperation } from "../operation"
-import { type Store, withdraw } from "../store"
-import { dispatchOrDeferStateUpdate } from "./dispatch-state-update"
-import { resetAtomOrSelector } from "./reset-atom-or-selector"
-import { RESET_STATE } from "./reset-in-store"
-import { setAtomOrSelector } from "./set-atom-or-selector"
+import type { Store } from "../store"
+import { operateOnStore } from "./operate-on-store"
+import type { RESET_STATE } from "./reset-in-store"
 
 export function setIntoStore<T, New extends T>(
 	store: Store,
@@ -44,117 +37,4 @@ export function setIntoStore<T, New extends T>(
 		  ]
 ): void {
 	operateOnStore(store, true, ...params)
-}
-
-export function operateOnStore<T, New extends T>(
-	store: Store,
-	ownOp: boolean,
-	...params:
-		| [
-				token: WritableFamilyToken<T, Canonical>,
-				key: Canonical,
-				value: New | typeof RESET_STATE | ((oldValue: T) => New),
-		  ]
-		| [
-				token: WritableToken<T>,
-				value: New | typeof RESET_STATE | ((oldValue: T) => New),
-		  ]
-): void {
-	let existingToken: WritableToken<T> | undefined
-	let brandNewToken: WritableToken<T> | undefined
-	let token: WritableToken<T>
-	let family: WritableFamilyToken<T, Canonical> | null
-	let key: Canonical | null
-	let value: New | typeof RESET_STATE | ((oldValue: T) => New)
-	if (params.length === 2) {
-		token = params[0]
-		value = params[1]
-		if (token.family) {
-			// biome-ignore lint/style/noNonNullAssertion: this token belongs to a family
-			family = getFamilyOfToken(store, token)!
-			key = parseJson(token.family.subKey)
-			// token = findInStore(store, family, key)
-			existingToken = seekInStore(store, family, key)
-			if (!existingToken) {
-				brandNewToken = mintInStore(store, family, key)
-				token = brandNewToken
-			} else {
-				token = existingToken
-			}
-		}
-	} else {
-		family = params[0]
-		key = params[1]
-		value = params[2]
-		existingToken = seekInStore(store, family, key)
-		if (!existingToken) {
-			brandNewToken = mintInStore(store, family, key)
-			token = brandNewToken
-		} else {
-			token = existingToken
-		}
-	}
-
-	const action = value === RESET_STATE ? `reset` : `set`
-
-	let target: Store & { operation: OpenOperation }
-
-	if (ownOp) {
-		const result = openOperation(store, token)
-		const rejected = typeof result === `number`
-		if (rejected) {
-			const rejectionTime = result
-			const unsubscribe = store.on.operationClose.subscribe(
-				`waiting to ${action} "${token.key}" at T-${rejectionTime}`,
-				function waitUntilOperationCloseToSetState() {
-					unsubscribe()
-					store.logger.info(
-						`🟢`,
-						token.type,
-						token.key,
-						`resuming deferred`,
-						action,
-						`from T-${rejectionTime}`,
-					)
-					operateOnStore(store, ownOp, token, value)
-				},
-			)
-			return
-		}
-		target = result
-	} else {
-		target = store as Store & { operation: OpenOperation }
-	}
-
-	if (`counterfeit` in token && `family` in token) {
-		const subKey = token.family.subKey
-		const disposal = store.disposalTraces.buffer.find(
-			(item) => item?.key === subKey,
-		)
-		store.logger.error(
-			`❌`,
-			token.type,
-			token.key,
-			`could not be`,
-			action,
-			`because it was not found in the store "${store.config.name}".`,
-			disposal
-				? `This state was previously disposed:\n${disposal.trace}`
-				: `No previous disposal trace was found.`,
-		)
-		return
-	}
-
-	const state = withdraw(target, token)
-	let protoUpdate: [T, T]
-	if (value === RESET_STATE) {
-		protoUpdate = resetAtomOrSelector(target, state)
-	} else {
-		protoUpdate = setAtomOrSelector(target, state, value)
-	}
-	dispatchOrDeferStateUpdate(target, state, protoUpdate, false)
-
-	if (ownOp) {
-		closeOperation(target)
-	}
 }
